requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
fastapi==0.93.0
ffmpy==0.3.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
gradio==3.20.1
gTTS==2.3.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.1
orjson==3.8.7
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycryptodome==3.17
pydantic==1.10.6
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2022.7.1
pywin32==305
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
yarl==1.8.2
platform
playsound
virtualenv
```
